@@ -17,17 +17,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "10.2.3",
-    "@angular/cdk": "10.2.7",
-    "@angular/common": "10.2.3",
-    "@angular/compiler": "10.2.3",
-    "@angular/core": "10.2.3",
-    "@angular/forms": "10.2.3",
-    "@angular/localize": "10.2.3",
-    "@angular/material": "10.2.7",
-    "@angular/platform-browser": "10.2.3",
-    "@angular/platform-browser-dynamic": "10.2.3",
-    "@angular/router": "10.2.3",
+    "@angular/animations": "11.0.1",
+    "@angular/cdk": "11.0.0",
+    "@angular/common": "11.0.1",
+    "@angular/compiler": "11.0.1",
+    "@angular/core": "11.0.1",
+    "@angular/forms": "11.0.1",
+    "@angular/localize": "11.0.1",
+    "@angular/material": "11.0.0",
+    "@angular/platform-browser": "11.0.1",
+    "@angular/platform-browser-dynamic": "11.0.1",
+    "@angular/router": "11.0.1",
     "@auth0/angular-jwt": "5.0.1",
     "@fortawesome/angular-fontawesome": "0.8.0",
     "@fortawesome/fontawesome-svg-core": "1.2.32",
@@ -51,31 +51,30 @@
     "ngx-filepond": "5.0.1",
     "normalizr": "3.6.1",
     "rxjs": "6.6.3",
-    "socket.io-client": "3.0.2",
-    "sweetalert2": "10.4.0",
-    "tslib": "2.0.3",
+    "socket.io-client": "3.0.3",
+    "tslib": "^2.0.0",
     "ua-parser-js": "0.7.22",
     "zone.js": "0.11.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "0.1100.1",
-    "@angular/cli": "10.2.0",
-    "@angular/compiler-cli": "10.2.3",
-    "@angular/language-service": "10.2.3",
+    "@angular-devkit/build-angular": "0.1100.2",
+    "@angular/cli": "11.0.2",
+    "@angular/compiler-cli": "11.0.1",
+    "@angular/language-service": "11.0.1",
     "@ngrx/schematics": "10.0.1",
-    "@types/jasmine": "3.6.1",
+    "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "2.0.8",
     "@types/lodash": "^4.14.165",
     "@types/node": "13.13.30",
     "@types/ua-parser-js": "0.7.33",
-    "codelyzer": "6.0.1",
-    "jasmine-core": "3.6.0",
-    "jasmine-spec-reporter": "6.0.0",
+    "codelyzer": "^6.0.0",
+    "jasmine-core": "~3.6.0",
+    "jasmine-spec-reporter": "~5.0.0",
     "karma": "5.2.3",
-    "karma-chrome-launcher": "3.1.0",
-    "karma-coverage-istanbul-reporter": "3.0.3",
-    "karma-jasmine": "4.0.1",
-    "karma-jasmine-html-reporter": "1.5.4",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage": "^2.0.3",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "^1.5.0",
     "prettier": "2.1.2",
     "protractor": "7.0.0",
     "ts-node": "9.0.0",
