@@ -1,17 +1,16 @@
 import { ChangeDetectorRef, Component, Input, OnDestroy } from '@angular/core';
-import { MatDialog } from '@angular/material/dialog';
 
 import { Store } from '@ngrx/store';
 import { Subject } from 'rxjs';
-import { delay, skipUntil, takeUntil } from 'rxjs/operators';
+import { takeUntil } from 'rxjs/operators';
 
 import { AppState } from '@/core/reducers';
-import { Asset } from '@/shared/models';
+import { Asset, DialogDataModel, DialogResultModel, DialogType, DialogWithTwoButtonModel } from '@/shared/models';
 import { TemplatesViewDetailsComponent } from '@/shared/components/templates';
-import { MediaqueryService, ToastMessageService } from '@/shared/services';
+import { DialogService, ToastMessageService } from '@/shared/services';
+import { CONSTANTS } from '@/shared/constants';
 import { Plan } from '@/modules/planner/models';
 import { User, UsersDialogComponent } from '@/modules/users';
-import { CONSTANTS } from '@/shared/constants';
 
 @Component({
   selector: 'templates-plan-view',
@@ -22,6 +21,7 @@ export class TemplatesPlanViewComponent extends TemplatesViewDetailsComponent<Pl
   @Input() canAddAttachment: boolean;
   @Input() canDeleteAttachment: boolean;
 
+  // TODO: @IMalaniak implement deleting document
   // @Output() deleteDocument: EventEmitter<any> = new EventEmitter();
 
   private unsubscribe: Subject<void> = new Subject();
@@ -29,9 +29,8 @@ export class TemplatesPlanViewComponent extends TemplatesViewDetailsComponent<Pl
   constructor(
     protected store$: Store<AppState>,
     protected toastMessageService: ToastMessageService,
-    private dialog: MatDialog,
-    private mediaQuery: MediaqueryService,
-    private cdr: ChangeDetectorRef
+    private cdr: ChangeDetectorRef,
+    private dialogService: DialogService
   ) {
     super(store$, toastMessageService);
   }
@@ -44,39 +43,27 @@ export class TemplatesPlanViewComponent extends TemplatesViewDetailsComponent<Pl
   }
 
   addParticipantDialog(): void {
-    const dialogRef = this.dialog.open(UsersDialogComponent, {
-      ...this.mediaQuery.deFaultPopupSize,
-      data: {
-        title: 'Select participants'
-      }
-    });
+    const dialogDataModel: DialogDataModel<DialogWithTwoButtonModel> = {
+      dialogModel: new DialogWithTwoButtonModel(CONSTANTS.TEXTS_SELECT_PARTICIPANS)
+    };
 
-    const userC = dialogRef.componentInstance.usersComponent;
-
-    dialogRef
-      .afterOpened()
-      .pipe(takeUntil(this.unsubscribe), skipUntil(userC.loading$), delay(300))
-      .subscribe(() => {
-        userC.users
-          .filter((user) => this.item.Participants.some((participant) => participant.id === user.id))
-          ?.forEach((selectedParticipant) => {
-            userC.selection.select(selectedParticipant);
-          });
-      });
-
-    dialogRef
+    this.dialogService
+      .open(UsersDialogComponent, dialogDataModel, DialogType.MAX)
       .afterClosed()
       .pipe(takeUntil(this.unsubscribe))
-      .subscribe((result: User[]) => {
-        const selectedParticipants: User[] = result?.filter(
-          (selectedParticipant) => !this.item.Participants.some((user) => user.id === selectedParticipant.id)
-        );
-
-        if (selectedParticipants?.length) {
-          this.item.Participants = [...this.item.Participants, ...selectedParticipants];
-          this.cdr.detectChanges();
+      .subscribe((result: DialogResultModel<User[]>) => {
+        if (result && result.success) {
+          const selectedParticipants: User[] = result.model.filter(
+            (selectedParticipant) => !this.item.Participants.some((user) => user.id === selectedParticipant.id)
+          );
+          if (selectedParticipants?.length) {
+            this.item.Participants = [...this.item.Participants, ...selectedParticipants];
+            this.cdr.detectChanges();
+          }
         }
       });
+
+    // TODO: @ArsenIrod add afterOpened logic
   }
 
   addDocument(doc: Asset): void {
