@@ -1,14 +1,14 @@
 import { ChangeDetectionStrategy, ChangeDetectorRef, Component, OnDestroy } from '@angular/core';
-import { MatDialog } from '@angular/material/dialog';
 
 import { Store } from '@ngrx/store';
 import { Subject } from 'rxjs';
-import { takeUntil, skipUntil, delay } from 'rxjs/operators';
+import { takeUntil } from 'rxjs/operators';
 
 import { AppState } from '@/core/reducers';
 import { TemplatesViewDetailsComponent } from '@/shared/components';
 import { MAT_BUTTON, COLUMN_NAMES, COLUMN_LABELS, CONSTANTS } from '@/shared/constants';
-import { MediaqueryService, ToastMessageService } from '@/shared/services';
+import { DialogService, ToastMessageService } from '@/shared/services';
+import { DialogDataModel, DialogResultModel, DialogType, DialogWithTwoButtonModel } from '@/shared/models';
 import { Privilege, Role } from '@/modules/roles/models';
 import { UsersDialogComponent, User } from '@/modules/users';
 import { PrivilegesDialogComponent } from '@/modules/roles/components/privileges/dialog/privileges-dialog.component';
@@ -36,95 +36,70 @@ export class TemplatesRoleViewComponent extends TemplatesViewDetailsComponent<Ro
   constructor(
     protected store$: Store<AppState>,
     protected toastMessageService: ToastMessageService,
-    private dialog: MatDialog,
-    private mediaQuery: MediaqueryService,
-    private cdr: ChangeDetectorRef
+    private cdr: ChangeDetectorRef,
+    private dialogService: DialogService
   ) {
     super(store$, toastMessageService);
   }
 
   addParticipantDialog(): void {
-    const dialogRef = this.dialog.open(UsersDialogComponent, {
-      ...this.mediaQuery.deFaultPopupSize,
-      data: {
-        title: 'Select Users'
-      }
-    });
+    const dialogDataModel: DialogDataModel<DialogWithTwoButtonModel> = {
+      dialogModel: new DialogWithTwoButtonModel(CONSTANTS.TEXTS_SELECT_USERS)
+    };
 
-    const userC = dialogRef.componentInstance.usersComponent;
-
-    dialogRef
-      .afterOpened()
-      .pipe(takeUntil(this.unsubscribe), skipUntil(userC.loading$), delay(300))
-      .subscribe(() => {
-        userC.users
-          .filter((user) => this.item.Users.some((rUser) => rUser.id === user.id))
-          ?.forEach((selectedParticipant) => {
-            userC.selection.select(selectedParticipant);
-          });
-      });
-
-    dialogRef
+    this.dialogService
+      .open(UsersDialogComponent, dialogDataModel, DialogType.MAX)
       .afterClosed()
       .pipe(takeUntil(this.unsubscribe))
-      .subscribe((result: User[]) => {
-        const selectedUsers: User[] = result?.filter(
-          (selectedUser) => !this.item.Users.some((user) => user.id === selectedUser.id)
-        );
-
-        if (selectedUsers?.length) {
-          this.item.Users = [...this.item.Users, ...selectedUsers];
-          this.cdr.detectChanges();
+      .subscribe((result: DialogResultModel<User[]>) => {
+        if (result && result.success) {
+          const selectedUsers: User[] = result.model.filter(
+            (selectedUser) => !this.item.Users.some((user) => user.id === selectedUser.id)
+          );
+          if (selectedUsers?.length) {
+            this.item.Users = [...this.item.Users, ...selectedUsers];
+            this.cdr.detectChanges();
+          }
         }
       });
+
+    // TODO: @ArsenIrod add afterOpened logic
   }
 
   addPrivilegeDialog(): void {
-    const dialogRef = this.dialog.open(PrivilegesDialogComponent, {
-      ...this.mediaQuery.deFaultPopupSize,
-      data: {
-        title: 'Select privileges'
-      }
-    });
+    const dialogDataModel: DialogDataModel<DialogWithTwoButtonModel> = {
+      dialogModel: new DialogWithTwoButtonModel(CONSTANTS.TEXTS_SELECT_PRIVILEGES)
+    };
 
-    const privilegesC = dialogRef.componentInstance.privilegesComponent;
-
-    dialogRef
-      .afterOpened()
-      .pipe(takeUntil(this.unsubscribe), skipUntil(privilegesC.isLoading$), delay(300))
-      .subscribe(() => {
-        privilegesC.privileges
-          .filter((privilege) => this.item.Privileges.some((rPrivilege) => rPrivilege.id === privilege.id))
-          ?.forEach((selectedPrivilege) => {
-            privilegesC.selection.select(selectedPrivilege);
-          });
-      });
-
-    dialogRef
+    this.dialogService
+      .open(PrivilegesDialogComponent, dialogDataModel, DialogType.STANDART)
       .afterClosed()
       .pipe(takeUntil(this.unsubscribe))
-      .subscribe((result: Privilege[]) => {
-        const selectedPrivileges: Privilege[] = result
-          ?.filter(
-            (selectedPrivilege) => !this.item.Privileges.some((rPrivilege) => rPrivilege.id === selectedPrivilege.id)
-          )
-          ?.map((selectedPrivilege) => {
-            return {
-              ...selectedPrivilege,
-              RolePrivilege: {
-                add: false,
-                view: false,
-                edit: false,
-                delete: false
-              }
-            };
-          });
-
-        if (selectedPrivileges?.length) {
-          this.item.Privileges = [...this.item.Privileges, ...selectedPrivileges];
-          this.cdr.detectChanges();
+      .subscribe((result: DialogResultModel<Privilege[]>) => {
+        if (result && result.success) {
+          const selectedPrivileges: Privilege[] = result.model
+            .filter(
+              (selectedPrivilege) => !this.item.Privileges.some((privilege) => privilege.id === selectedPrivilege.id)
+            )
+            ?.map((selectedPrivilege) => {
+              return {
+                ...selectedPrivilege,
+                RolePrivilege: {
+                  add: false,
+                  view: false,
+                  edit: false,
+                  delete: false
+                }
+              };
+            });
+          if (selectedPrivileges?.length) {
+            this.item.Privileges = [...this.item.Privileges, ...selectedPrivileges];
+            this.cdr.detectChanges();
+          }
         }
       });
+
+    // TODO: @ArsenIrod add afterOpened logic
   }
 
   removePrivilege(privilegeId: number): void {
