@@ -1,14 +1,21 @@
 import { waitForAsync, ComponentFixture, TestBed } from '@angular/core/testing';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { MAT_DIALOG_DATA } from '@angular/material/dialog';
+import { RouterTestingModule } from '@angular/router/testing';
+import { HttpClientModule } from '@angular/common/http';
 
 import { provideMockStore } from '@ngrx/store/testing';
 
 import { SharedModule } from '@/shared/shared.module';
 import { formsStateMock } from '@/shared/testing/mocks';
 import { OrganismsTaskDialogComponent } from './organisms-task-dialog.component';
-import { HttpClientModule } from '@angular/common/http';
+import { DialogCreateEditModel, DialogDataModel, DialogMode } from '@/shared/models/dialog';
 
 describe('OrganismsTaskDialogComponent', () => {
+  const dialogDataModel: DialogDataModel<DialogCreateEditModel> = {
+    dialogModel: new DialogCreateEditModel(DialogMode.CREATE)
+  };
+
   let component: OrganismsTaskDialogComponent;
   let fixture: ComponentFixture<OrganismsTaskDialogComponent>;
   const initialState = {
@@ -19,8 +26,8 @@ describe('OrganismsTaskDialogComponent', () => {
     waitForAsync(() => {
       TestBed.configureTestingModule({
         declarations: [OrganismsTaskDialogComponent],
-        imports: [HttpClientModule, SharedModule, BrowserAnimationsModule],
-        providers: [provideMockStore({ initialState })]
+        imports: [HttpClientModule, SharedModule, BrowserAnimationsModule, RouterTestingModule],
+        providers: [provideMockStore({ initialState }), { provide: MAT_DIALOG_DATA, useValue: dialogDataModel }]
       }).compileComponents();
     })
   );
