@@ -1,18 +1,17 @@
 import { Component, OnInit, OnDestroy, ViewChild, ElementRef, ChangeDetectionStrategy } from '@angular/core';
 import { Router, NavigationEnd, RouterOutlet } from '@angular/router';
+import { Observable, Subject } from 'rxjs';
 import { filter, takeUntil } from 'rxjs/operators';
 import { Store, select } from '@ngrx/store';
 import { AppState } from '@/core/reducers';
-import { Observable, Subject } from 'rxjs';
 import { User } from '@/modules/users';
 import { currentUser } from '@/core/auth/store/auth.selectors';
 import * as layoutActions from '../store/layout.actions';
 import * as fromLayout from '../store';
-import { MediaqueryService } from '@/shared/services';
 import { privateRouterTransition } from '@/shared/animations';
+import { MediaQueryService } from '@/core/services';
 
 @Component({
-  selector: 'app-private',
   template: `
     <section class="grid" [ngClass]="{ 'dark-theme': enableDarkTheme$ | async, 'font-scale': scaleFontUp$ | async }">
       <header-component
@@ -27,7 +26,7 @@ import { privateRouterTransition } from '@/shared/animations';
         <section #contentWrapper class="content" [ngClass]="{ 'dark-theme-bg': enableDarkTheme$ | async }">
           <div
             class="overlay"
-            *ngIf="mediaquery.isMobileDevice"
+            *ngIf="mediaQueryService.isMobileDevice"
             [ngClass]="{ isVisible: !(leftSidebarMinimized$ | async) || !(rightSidebarMinimized$ | async) }"
             (click)="onOverlayClick()"
           ></div>
@@ -51,18 +50,18 @@ import { privateRouterTransition } from '@/shared/animations';
   changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class PrivateViewComponent implements OnInit, OnDestroy {
-  currentUser$: Observable<User> = this.store.pipe(select(currentUser));
-  scaleFontUp$: Observable<boolean> = this.store.pipe(select(fromLayout.getScalledFontState));
-  enableDarkTheme$: Observable<boolean> = this.store.pipe(select(fromLayout.getDarkThemeState));
-  leftSidebarMinimized$: Observable<boolean> = this.store.pipe(select(fromLayout.getLeftSidebarState));
-  rightSidebarMinimized$: Observable<boolean> = this.store.pipe(select(fromLayout.getRightSidebarState));
+  currentUser$: Observable<User> = this.store$.pipe(select(currentUser));
+  scaleFontUp$: Observable<boolean> = this.store$.pipe(select(fromLayout.getScalledFontState));
+  enableDarkTheme$: Observable<boolean> = this.store$.pipe(select(fromLayout.getDarkThemeState));
+  leftSidebarMinimized$: Observable<boolean> = this.store$.pipe(select(fromLayout.getLeftSidebarState));
+  rightSidebarMinimized$: Observable<boolean> = this.store$.pipe(select(fromLayout.getRightSidebarState));
 
   @ViewChild('contentWrapper', { static: false })
   contentWrapper: ElementRef;
 
   private unsubscribe: Subject<void> = new Subject();
 
-  constructor(private router: Router, public mediaquery: MediaqueryService, private store: Store<AppState>) {}
+  constructor(private router: Router, public mediaQueryService: MediaQueryService, private store$: Store<AppState>) {}
 
   ngOnInit(): void {
     this.router.events
@@ -71,7 +70,7 @@ export class PrivateViewComponent implements OnInit, OnDestroy {
         takeUntil(this.unsubscribe)
       )
       .subscribe(() => {
-        if (this.mediaquery.isMobileDevice) {
+        if (this.mediaQueryService.isMobileDevice) {
           this.toggleLeftSidebar(true);
           this.toggleRightSidebar(true);
         }
@@ -81,19 +80,19 @@ export class PrivateViewComponent implements OnInit, OnDestroy {
   }
 
   toggleLeftSidebar(minimized: boolean): void {
-    this.store.dispatch(layoutActions.toggleLeftSidebar({ minimized }));
+    this.store$.dispatch(layoutActions.toggleLeftSidebar({ minimized }));
   }
 
   toggleRightSidebar(minimized: boolean): void {
-    this.store.dispatch(layoutActions.toggleRightSidebar({ minimized }));
+    this.store$.dispatch(layoutActions.toggleRightSidebar({ minimized }));
   }
 
   enableDarkTheme(enabled: boolean): void {
-    this.store.dispatch(layoutActions.enableDarkTheme({ enabled }));
+    this.store$.dispatch(layoutActions.enableDarkTheme({ enabled }));
   }
 
   scaleFontUp(scaled: boolean): void {
-    this.store.dispatch(layoutActions.scaleFontUp({ scaled }));
+    this.store$.dispatch(layoutActions.scaleFontUp({ scaled }));
   }
 
   onOverlayClick(): void {
