@@ -1,33 +1,35 @@
 import { Component, Inject, ViewChild, ChangeDetectionStrategy } from '@angular/core';
 import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
-import { RolesComponent } from '../list/roles.component';
-import { Role } from '../../models/role';
-import { ACTION_LABELS, THEME_PALETTE } from '@/shared/constants';
+import { ComponentType } from '@angular/cdk/portal/portal';
 
-export interface RolesDialogData {
-  title: string;
-}
+import { RolesComponent } from '../list/roles.component';
+import { ACTION_LABELS, THEME_PALETTE } from '@/shared/constants';
+import { DialogDataModel, DialogWithTwoButtonModel, DialogResultModel } from '@/shared/models';
+import { DialogBaseModel } from '@/shared/components';
+import { Role } from '../../models';
+
 @Component({
   templateUrl: 'roles-dialog.component.html',
   changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class RolesDialogComponent {
+export class RolesDialogComponent extends DialogBaseModel<DialogWithTwoButtonModel> {
+  @ViewChild(RolesComponent, { static: true }) rolesComponent: RolesComponent;
+
   actionLabels = ACTION_LABELS;
   themePalette = THEME_PALETTE;
 
   constructor(
-    public dialogRef: MatDialogRef<RolesDialogComponent>,
-    @Inject(MAT_DIALOG_DATA) public data: RolesDialogData
-  ) {}
-
-  @ViewChild(RolesComponent, { static: true })
-  rolesComponent: RolesComponent;
-
-  onNoClick(): void {
-    this.dialogRef.close();
+    readonly dialogRef: MatDialogRef<ComponentType<unknown>>,
+    @Inject(MAT_DIALOG_DATA) protected data: DialogDataModel<DialogWithTwoButtonModel>
+  ) {
+    super(dialogRef, data);
   }
 
-  onSubmitClick(roles: Role[]): void {
-    this.dialogRef.close(roles);
+  onClose(success: boolean): void {
+    const result: DialogResultModel<Role[]> = {
+      success,
+      model: this.rolesComponent.selection.selected
+    };
+    return this.dialogRef.close(result);
   }
 }
