@@ -1,32 +1,34 @@
 import { Component, Inject, ViewChild } from '@angular/core';
 import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
+import { ComponentType } from '@angular/cdk/portal';
+
 import { PrivilegesComponent } from '../list/privileges.component';
 import { ACTION_LABELS, THEME_PALETTE } from '@/shared/constants';
-
-export interface PrivilegesDialogData {
-  title: string;
-}
+import { DialogDataModel, DialogWithTwoButtonModel, DialogResultModel } from '@/shared/models';
+import { DialogBaseModel } from '@/shared/components';
+import { Privilege } from '@/modules/roles/models';
 
 @Component({
   templateUrl: 'privileges-dialog.component.html'
 })
-export class PrivilegesDialogComponent {
+export class PrivilegesDialogComponent extends DialogBaseModel<DialogWithTwoButtonModel> {
+  @ViewChild(PrivilegesComponent, { static: true }) privilegesComponent: PrivilegesComponent;
+
   actionLabels = ACTION_LABELS;
   themePalette = THEME_PALETTE;
 
   constructor(
-    public dialogRef: MatDialogRef<PrivilegesDialogComponent>,
-    @Inject(MAT_DIALOG_DATA) public data: PrivilegesDialogData
-  ) {}
-
-  @ViewChild(PrivilegesComponent, { static: true })
-  privilegesComponent: PrivilegesComponent;
-
-  onNoClick(): void {
-    this.dialogRef.close();
+    readonly dialogRef: MatDialogRef<ComponentType<unknown>>,
+    @Inject(MAT_DIALOG_DATA) protected data: DialogDataModel<DialogWithTwoButtonModel>
+  ) {
+    super(dialogRef, data);
   }
 
-  onSubmitClick(): void {
-    this.dialogRef.close(this.privilegesComponent.selection.selected);
+  onClose(success: boolean): void {
+    const result: DialogResultModel<Privilege[]> = {
+      success,
+      model: this.privilegesComponent.selection.selected
+    };
+    this.dialogRef.close(result);
   }
 }
