@@ -1,23 +1,29 @@
 import { Component, OnInit, OnDestroy } from '@angular/core';
-import { MatDialog } from '@angular/material/dialog';
+import { SelectionModel } from '@angular/cdk/collections';
+import { takeUntil, map } from 'rxjs/operators';
+import { Subject, Observable } from 'rxjs';
+
+import { Store, select } from '@ngrx/store';
+
 import { Privilege } from '../../../models';
 import { AddPrivilegeDialogComponent } from '../add-dialog/add-privilege-dialog.component';
-import { Subject, Observable } from 'rxjs';
-import { Store, select } from '@ngrx/store';
 import { AppState } from '@/core/reducers';
 import { allPrivilegesRequested, createPrivilegeRequested } from '@/modules/roles/store/privilege.actions';
 import { selectAllPrivileges, selectPrivilegesLoading } from '@/modules/roles/store/privilege.selectors';
-import { takeUntil, map } from 'rxjs/operators';
-import { SelectionModel } from '@angular/cdk/collections';
-import { COLUMN_NAMES, COLUMN_LABELS, ACTION_LABELS } from '@/shared/constants';
+import { COLUMN_NAMES, COLUMN_LABELS, ACTION_LABELS, CONSTANTS } from '@/shared/constants';
+import { DialogDataModel } from '@/shared/models/dialog/dialog-data.model';
+import { DialogService } from '@/shared/services';
+import { DialogMode } from '@/shared/models/dialog/dialog-mode.enum';
+import { DialogResultModel } from '@/shared/models/dialog/dialog-result.model';
+import { DialogCreateEditModel, DialogType } from '@/shared/models';
 
 @Component({
-  selector: 'app-privileges',
+  selector: 'privileges-component',
   templateUrl: './privileges.component.html',
   styleUrls: ['./privileges.component.scss']
 })
 export class PrivilegesComponent implements OnInit, OnDestroy {
-  isLoading$: Observable<boolean> = this.store.pipe(select(selectPrivilegesLoading));
+  isLoading$: Observable<boolean> = this.store$.pipe(select(selectPrivilegesLoading));
 
   selection = new SelectionModel<Privilege>(true, []);
   privileges: Privilege[];
@@ -30,11 +36,11 @@ export class PrivilegesComponent implements OnInit, OnDestroy {
 
   private unsubscribe: Subject<void> = new Subject();
 
-  constructor(private dialog: MatDialog, private store: Store<AppState>) {}
+  constructor(private store$: Store<AppState>, private dialogService: DialogService) {}
 
   ngOnInit(): void {
-    this.store.dispatch(allPrivilegesRequested());
-    this.store
+    this.store$.dispatch(allPrivilegesRequested());
+    this.store$
       .pipe(
         takeUntil(this.unsubscribe),
         select(selectAllPrivileges),
@@ -60,14 +66,20 @@ export class PrivilegesComponent implements OnInit, OnDestroy {
   }
 
   createPrivilegeDialog(): void {
-    const dialogRef = this.dialog.open(AddPrivilegeDialogComponent);
+    const dialogModel = new DialogCreateEditModel(
+      DialogMode.CREATE,
+      CONSTANTS.TEXTS_CREATE_PRIVILEGE,
+      ACTION_LABELS.SUMBIT
+    );
+    const dialogDataModel: DialogDataModel<DialogCreateEditModel> = { dialogModel };
 
-    dialogRef
+    this.dialogService
+      .open(AddPrivilegeDialogComponent, dialogDataModel, DialogType.STANDART)
       .afterClosed()
       .pipe(takeUntil(this.unsubscribe))
-      .subscribe((privilege: Privilege) => {
-        if (privilege) {
-          this.store.dispatch(createPrivilegeRequested({ privilege }));
+      .subscribe((result: DialogResultModel<Privilege>) => {
+        if (result && result.success) {
+          this.store$.dispatch(createPrivilegeRequested({ privilege: result.model }));
         }
       });
   }
