@@ -1,5 +1,6 @@
 import { Component, OnInit, Inject, ChangeDetectionStrategy } from '@angular/core';
 import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
+import { ComponentType } from '@angular/cdk/portal';
 
 import { select, Store } from '@ngrx/store';
 import { Observable } from 'rxjs';
@@ -9,15 +10,16 @@ import { selectFormByName } from '@/core/reducers/dynamic-form/dynamic-form.sele
 import { formRequested } from '@/core/reducers/dynamic-form/dynamic-form.actions';
 import { DynamicForm } from '@/shared/models';
 import { ACTION_LABELS, FORMCONSTANTS, THEME_PALETTE } from '@/shared/constants';
-import { Task, TaskDialogData } from '@/modules/task-manager/models';
+import { DialogCreateEditPageModel } from '@/shared/components';
+import { DialogCreateEditModel, DialogDataModel, DialogResultModel } from '@/shared/models';
+import { Task } from '@/modules/task-manager/models';
 
 @Component({
   selector: 'organisms-task-dialog',
   templateUrl: './organisms-task-dialog.component.html',
-  styleUrls: ['./organisms-task-dialog.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class OrganismsTaskDialogComponent implements OnInit {
+export class OrganismsTaskDialogComponent extends DialogCreateEditPageModel implements OnInit {
   taskFormJson$: Observable<DynamicForm> = this.store$.pipe(select(selectFormByName(FORMCONSTANTS.TASK)));
 
   taskFormValues: Task;
@@ -27,9 +29,11 @@ export class OrganismsTaskDialogComponent implements OnInit {
 
   constructor(
     private store$: Store<AppState>,
-    public dialogRef: MatDialogRef<OrganismsTaskDialogComponent>,
-    @Inject(MAT_DIALOG_DATA) public data: TaskDialogData
-  ) {}
+    readonly dialogRef: MatDialogRef<ComponentType<Task>>,
+    @Inject(MAT_DIALOG_DATA) protected data: DialogDataModel<DialogCreateEditModel>
+  ) {
+    super(dialogRef, data);
+  }
 
   ngOnInit(): void {
     this.store$.dispatch(formRequested({ formName: FORMCONSTANTS.TASK }));
@@ -39,7 +43,11 @@ export class OrganismsTaskDialogComponent implements OnInit {
     this.taskFormValues = { ...this.taskFormValues, ...formVal };
   }
 
-  onSubmit(): void {
-    this.dialogRef.close({ ...this.data.task, ...this.taskFormValues });
+  onClose(success: boolean): void {
+    const result: DialogResultModel<Task> = {
+      success,
+      model: { ...this.model, ...this.taskFormValues }
+    };
+    this.dialogRef.close(result);
   }
 }
