@@ -2,7 +2,7 @@ import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { Task, TaskPriority } from '../models';
 import { Observable } from 'rxjs';
-import { ApiResponse, CollectionApiResponse, ItemApiResponse } from '@/shared/models';
+import { BaseMessage, CollectionApiResponse, ItemApiResponse } from '@/shared/models';
 import { APIS } from '@/shared/constants';
 
 @Injectable({
@@ -19,11 +19,11 @@ export class TaskService {
     return this.http.put<ItemApiResponse<Task>>(`${APIS.TASKS}/${task.id}`, task);
   }
 
-  delete(id: number): Observable<ApiResponse> {
-    return this.http.delete<ApiResponse>(`${APIS.TASKS}/${id}`);
+  delete(id: number): Observable<BaseMessage> {
+    return this.http.delete<BaseMessage>(`${APIS.TASKS}/${id}`);
   }
 
-  deleteMultipleTask(taskIds: number[]): Observable<ApiResponse> {
+  deleteMultipleTask(taskIds: number[]): Observable<BaseMessage> {
     // TODO @ArseniiIrod, @IMalaniak change this to delete request with body
     return this.http.put<any>(`${APIS.TASKS_MULTIPLE}/${1}`, { taskIds });
   }
