@@ -1,7 +1,11 @@
-export enum DIALOG {
-  CONFIRM = 'Are You sure?',
-  OK = 'OK',
-  CANCEL = 'Cancel',
-  YES = 'Yes',
-  NO = 'No'
+export class DIALOG {
+  static CONFIRM = 'Are You sure?';
+  static OK = 'OK';
+  static CANCEL = 'Cancel';
+  static YES = 'Yes';
+  static NO = 'No';
+  static SAVE = 'Save';
+  static CREATE = 'Create';
+  static MAX_WIDTH = '90vw';
+  static MAX_HEIGHT = '90vh';
 }
