@@ -30,6 +30,11 @@ const TEMPLATES = [
   _components.TemplatesBoxUserListSmComponent,
   _components.TemplatesUserDetailsComponent
 ];
+const DIALOGS = [
+  _components.DialogBaseComponent,
+  _components.DialogConfirmComponent,
+  _components.DialogWithTwoButtonsComponent
+];
 
-export const COMPONENTS = [...ATOMS, ...MOLECULES, ...ORGANISMS, ...TEMPLATES];
+export const COMPONENTS = [...ATOMS, ...MOLECULES, ...ORGANISMS, ...TEMPLATES, ...DIALOGS];
 export const PIPES = [_pipes.MemoizePipe, _pipes.DateTimeFormatPipe];
