@@ -1,6 +1,6 @@
-﻿export * from './common.datasource';
-export * from './dynamic-form.resolver';
-export * from './mediaquery.service';
+﻿export * from './dialog.service';
+export * from './common.datasource';
 export * from './local-storage.service';
 export * from './socket.service';
 export * from './toast-message.service';
+export * from './dynamic-form.resolver';
